refactor(svelte-app): extract SavedAccounts component and share shortenNpub
DRY + componentization per review feedback:
- Add shortenNpub() to bech32-converter (hex→shortened npub with hex fallback,
  configurable head/tail) and reuse it in the account list, PublicKeyDisplay,
  and ConsentDialog, which each had their own near-identical truncation.
- Move the saved-accounts list (re-login + delete + countdown + animations)
  out of AuthScreen into a self-contained SavedAccounts.svelte that owns its
  own state and reports re-login errors via an onError callback. AuthScreen
  now just renders <SavedAccounts> in the login tab.
- Add unit tests for shortenNpub (default, custom head/tail, fallback).
